Move test_utils.go content to helpers_test.go
