--- lib/trie/test_utils.go
@@ -1,75 +1,0 @@
-// Copyright 2021 ChainSafe Systems (ON)
-// SPDX-License-Identifier: LGPL-3.0-only
-
-package trie
-
-import (
-	"math/rand"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-// Test represents a key-value pair for a test
-type Test struct {
-	key   []byte
-	value []byte
-	op    int
-}
-
-func generateKeyValues(tb testing.TB, seed int64, size int) (kv map[string][]byte) {
-	tb.Helper()
-
-	kv = make(map[string][]byte, size)
-
-	generator := rand.New(rand.NewSource(seed)) //nolint:gosec
-
-	const maxKeySize, maxValueSize = 510, 128
-	for i := 0; i < size; i++ {
-		populateKeyValueMap(tb, kv, generator, maxKeySize, maxValueSize)
-	}
-
-	return kv
-}
-
-func populateKeyValueMap(tb testing.TB, kv map[string][]byte,
-	generator *rand.Rand, maxKeySize, maxValueSize int) {
-	tb.Helper()
-
-	for {
-		const minKeySize = 2
-		key := generateRandBytesMinMax(tb, minKeySize, maxKeySize, generator)
-
-		keyString := string(key)
-
-		_, keyExists := kv[keyString]
-
-		if keyExists && key[1] != byte(0) {
-			continue
-		}
-
-		const minValueSize = 2
-		value := generateRandBytesMinMax(tb, minValueSize, maxValueSize, generator)
-
-		kv[keyString] = value
-
-		break
-	}
-}
-
-func generateRandBytesMinMax(tb testing.TB, minSize, maxSize int,
-	generator *rand.Rand) (b []byte) {
-	tb.Helper()
-	size := minSize +
-		generator.Intn(maxSize-minSize)
-	return generateRandBytes(tb, size, generator)
-}
-
-func generateRandBytes(tb testing.TB, size int,
-	generator *rand.Rand) (b []byte) {
-	tb.Helper()
-	b = make([]byte, size)
-	_, err := generator.Read(b)
-	require.NoError(tb, err)
-	return b
-}
